pyproject.toml
[project]
name = "riff-mcp"
version = "0.1.0"
description = "Batch Gemini media generation CLI plus companion MCP servers (gemini-prompts-mcp, media-analysis-mcp) for the riff workflow — generate, analyze, iterate."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.47.0",
    "pillow>=10.4.0",
    "PyYAML>=6.0.2",
    "python-dotenv>=1.0.1",
    "mcp[cli]>=1.0.0",
    "replicate>=0.34.0",
    "pydantic>=2.0",
]

[project.scripts]
gemini-video-prompts = "gemini_video_prompts.cli:main"
gemini-prompts-mcp = "gemini_video_prompts_mcp.server:main"
media-analysis-mcp = "media_analysis_mcp.server:main"
riff-mcp-doctor = "riff_mcp_doctor.doctor:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
  "src/gemini_video_prompts",
  "src/gemini_video_prompts_mcp",
  "src/media_analysis_mcp",
  "src/riff_mcp_doctor",
]

[tool.hatch.build.targets.sdist]
include = [
  "src/gemini_video_prompts",
  "src/gemini_video_prompts_mcp",
  "src/media_analysis_mcp",
  "src/riff_mcp_doctor",
  "pyproject.toml",
  "README.md",
  ".env.example",
  ".mcp.example.json",
  "prompts",
]

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]
